+++ src/main/java/basemod/ModSlider.java
@@ -12,14 +12,14 @@ import com.megacrit.cardcrawl.helpers.input.InputHelper;
 import java.util.function.Consumer;
 
 public class ModSlider implements IUIElement {
-    private static final float BG_X = 1350.0f * Settings.scale;
     private static final float L_X = 1235.0f * Settings.scale;
     private static final float SLIDE_W = 230.0f * Settings.scale;
     
     private Consumer<ModSlider> change;
     private Hitbox hb;
     private Hitbox bgHb;
-    private float x;
+    private float sliderX;
+    private float handleX;
     private float y;
     private boolean sliderGrabbed = false;
     private String label;
@@ -42,27 +42,48 @@ public class ModSlider implements IUIElement {
         } else {
             posX *= Settings.scale;
         }
-        x = posX + (SLIDE_W * value);
+        sliderX = posX - 11.0f * Settings.scale;
+        handleX = posX + SLIDE_W;
         y = posY * Settings.scale;
         
         hb = new Hitbox(42.0f * Settings.scale, 38.0f * Settings.scale);
         bgHb = new Hitbox(300.0f * Settings.scale, 38.0f * Settings.scale);
-        bgHb.move(BG_X, y);
+        bgHb.move(sliderX + SLIDE_W / 2.0f, y);
     }
     
     public void render(SpriteBatch sb) {
         sb.setColor(Color.WHITE);
         
-        sb.draw(ImageMaster.OPTION_SLIDER_BG, BG_X - 125.0f, y - 12.0f, 125.0f, 12.0f, 250.0f, 24.0f, Settings.scale, Settings.scale, 0.0f, 0, 0, 250, 24, false, false);
-        sb.draw(ImageMaster.OPTION_SLIDER, x - 22.0f, y - 22.0f, 22.0f, 22.0f, 44.0f, 44.0f, Settings.scale, Settings.scale, 0.0f, 0, 0, 44, 44, false, false);
+        sb.draw(
+                ImageMaster.OPTION_SLIDER_BG,
+                sliderX, y - 12.0f,
+                0.0f, 12.0f,
+                250.0f, 24.0f,
+                Settings.scale, Settings.scale,
+                0.0f,
+                0, 0,
+                250, 24,
+                false, false
+        );
+        sb.draw(
+                ImageMaster.OPTION_SLIDER,
+                handleX - 22.0f, y - 22.0f,
+                22.0f, 22.0f,
+                44.0f, 44.0f,
+                Settings.scale, Settings.scale,
+                0.0f,
+                0, 0,
+                44, 44,
+                false, false
+        );
         
-        FontHelper.renderFontCentered(sb, FontHelper.tipBodyFont, label, BG_X - 170.0f * Settings.scale, y, Color.WHITE);
+        FontHelper.renderFontCentered(sb, FontHelper.tipBodyFont, label, sliderX - 55.0f * Settings.scale, y, Color.WHITE);
         
         String renderVal = Integer.toString(Math.round(value * multiplier));
         if (sliderGrabbed) {
-            FontHelper.renderFontCentered(sb, FontHelper.tipBodyFont, renderVal + suffix, BG_X + 170.0f * Settings.scale, y, Settings.GREEN_TEXT_COLOR);
+            FontHelper.renderFontCentered(sb, FontHelper.tipBodyFont, renderVal + suffix, sliderX + SLIDE_W + 55.0f * Settings.scale, y, Settings.GREEN_TEXT_COLOR);
         } else {
-            FontHelper.renderFontCentered(sb, FontHelper.tipBodyFont, renderVal + suffix, BG_X + 170.0f * Settings.scale, y, Settings.BLUE_TEXT_COLOR);
+            FontHelper.renderFontCentered(sb, FontHelper.tipBodyFont, renderVal + suffix, sliderX + SLIDE_W + 55.0f * Settings.scale, y, Settings.BLUE_TEXT_COLOR);
         }
         
         this.hb.render(sb);
@@ -72,12 +93,12 @@ public class ModSlider implements IUIElement {
     public void update() {        
         hb.update();
         bgHb.update();
-        hb.move(x, y);
+        hb.move(handleX, y);
         
         if (sliderGrabbed) {    
             if (InputHelper.isMouseDown) {
-                x = MathHelper.fadeLerpSnap(x, InputHelper.mX);
-                x = Math.min(L_X + SLIDE_W, Math.max(x, L_X));
+                handleX = MathHelper.fadeLerpSnap(handleX, InputHelper.mX);
+                handleX = Math.min(sliderX + SLIDE_W + 11.0f * Settings.scale, Math.max(handleX, sliderX + 11.0f * Settings.scale));
             } else {
                 sliderGrabbed = false;
             }
@@ -88,7 +109,7 @@ public class ModSlider implements IUIElement {
         }
         
         int oldVal = Math.round(value * multiplier);
-        value = (x - L_X) / SLIDE_W;
+        value = (handleX - 11.0f * Settings.scale - sliderX) / SLIDE_W;
         
         if (oldVal != Math.round(value * multiplier)) {
             onChange();
@@ -96,7 +117,7 @@ public class ModSlider implements IUIElement {
     }
     
     public void setValue(float val) {
-        x = L_X + (SLIDE_W * val);
+        handleX = sliderX + (SLIDE_W * val) + 11.0f * Settings.scale;
     }
     
     private void onChange() {