+++ CHANGELOG.md
@@ -440,3 +440,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 #### dev ####
 * Fix custom monsters being more weighted (kiooeht)
 * Fix custom monsters not being used for the first strong enemy (kiooeht)
+* More fixing ModSlider position (kiooeht)