+++ CHANGELOG.md
@@ -437,8 +437,10 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix Discovery not obeying new max hand size (kiooeht)
 * Fix ModSlider using y position as x position (kiooeht)
 
-#### dev ####
+#### v5.0.0 ####
 * Support for week 46 (kiooeht)
 * Fix custom monsters being more weighted (kiooeht)
 * Fix custom monsters not being used for the first strong enemy (kiooeht)
 * More fixing ModSlider position (kiooeht)
+
+#### dev ####